requirements.txt
openai==2.21.0
boto3
python-dotenv==1.2.1